isolated publisher logic
